% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Add.Spatial.Expr.Pattern}
\alias{Add.Spatial.Expr.Pattern}
\title{Adds spatial differential expressed pattern (region-specific effects) to
a cell type.}
\usage{
Add.Spatial.Expr.Pattern(
  sim.count,
  r,
  CellType,
  GeneID = NULL,
  PropOfGenes = 0.1,
  delta.mean = 1,
  delta.sd = 0.01,
  seed
)
}
\arguments{
\item{sim.count}{Simulated expression counts from single-cell expression
data, before adding in additional spatial patterns.}

\item{r}{Which region to add in the spatial pattern.}

\item{CellType}{Which cell type to add in the spatial pattern.}

\item{GeneID}{Gene(s) index. Default = NULL, and then a random subset
of genes will be perturbed based on the defined spatial patterns.}

\item{PropOfGenes}{Proportion of genes with this pattern if GeneID is not
provided.}

\item{delta.mean}{Expected effect size (at the log scale of the counts).}

\item{delta.sd}{Standard deviation of the effect size.}

\item{seed}{Seed}
}
\value{
\item{SignalSummary:}{Summary of this spatial pattern, including the type
of spatial patterns, impacted cell types, perturbed genes, and effect sizes.}
\item{beta.matrix:}{Effect size on each gene in each cell. }
}
\description{
This function adds one type of spatial differential expressed patterns. This
function can be repeated used to add region-specific effects  in different
regions for different cell types.
}
