% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_STData.R
\name{cell.loc.existing.fc}
\alias{cell.loc.existing.fc}
\title{Generate cell location data by using existing SRT data directly.}
\usage{
cell.loc.existing.fc(
  PointLoc,
  PointAnno,
  PointRegion,
  window_method = "rectangle"
)
}
\arguments{
\item{PointLoc}{The location of input cells on x, y axis.}

\item{PointAnno}{The cell type annotation of input cells.}

\item{PointRegion}{The spatial regions of input cells.}

\item{window_method}{Method for estimating window of cells. Choices include "network", "convex", "convex2",
"convex3", convex5", "rectangle". Default is "network", which is most accurate.}
}
\value{
Simulated cell spatial map file, which is the same as the reference spatial data.
}
\description{
Generate cell location data by using existing SRT data directly.
}
