% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_NoData.R
\name{connectUp}
\alias{connectUp}
\title{Assign connected regions in a window}
\usage{
connectUp(r, nRegion, seed = NULL)
}
\arguments{
\item{r}{polygon generated in \code{RandomRegionWindow} function.}

\item{nRegion}{nRegion is the No. of regions (e.g. nRegion=3)}

\item{seed}{Random seed.}
}
\value{
A list of the selected polygons for each region.
}
\description{
This function assigns random connected regions on a square. Used within function \code{RandomRegionWindow}.
}
