% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{df_to_config}
\alias{df_to_config}
\title{Convert a data.frame to config.yml}
\usage{
df_to_config(x, parameter_file = "config.yml")
}
\arguments{
\item{x}{a data.frame with two columns (parameters and value)}

\item{parameter_file}{name for the config file. Default = config.yml}
}
\value{
a config.yml file
}
\description{
Convert a data.frame to config.yml
}
\examples{
\dontrun{
df <- data.frame(
    parameters = c("expression_data_file", "num_simulated_cells"),
    value = c("fake1_expr.Rdata", 10000))
df_to_config(df)
}
}
