% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Est_ModelPara}
\alias{Est_ModelPara}
\title{Estimate Marginal distribution and Gaussian Copula for Gene Expression Matrix}
\usage{
Est_ModelPara(
  expr,
  anno,
  sim_method = c("ind", "copula"),
  region = NULL,
  ncores = 1,
  subsetN = NULL
)
}
\arguments{
\item{expr}{Expression levels of input data}

\item{anno}{Cell type annotation of input data}

\item{sim_method}{c('ind', 'copula')}

\item{region}{description}

\item{ncores}{No of cores for parallel computing.}

\item{subsetN}{If the number of cells is large, one can use it to estimate expression distribuitons from subset of cells (e.g. subsetN=2500) per cell type.}
}
\value{
Estimated Gaussian Copula
}
\description{
Estimate Marginal distribution and Gaussian Copula for Gene Expression Matrix
}
