% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{EstCCI}
\alias{EstCCI}
\title{Estimate three patterns for cell-cell interactions}
\usage{
EstCCI(
  expr_data,
  spatial_data,
  paired = T,
  LR_database = c("mouse", "human", "external", "null"),
  LR_database_path = NULL,
  save_folder = getwd()
)
}
\arguments{
\item{expr_data}{Input expression data.}

\item{spatial_data}{Input spatial data.}

\item{paired}{If the expression and spatial data are paired. If not, analysis will not be run.}

\item{LR_database}{sCCIgen provides two ligand-recepter databases from CellTalkDB for mouse and human. Users
can select "external" and provide the path to external database at LR_database_path. Users can also select
"null", which will not run LR analysis.)}

\item{save_folder}{Provide a path to a folder that saves the CCI results for use in sCCIgen.}
}
\value{
No return inside R. Results are saved in \code{save_folder}.
File est_CCI_dist_dist.tsv will save results of CCI through cell location attraction and inhibition.
File est_CCI_expr_dist.tsv will save results of CCI through expression via cell location impact.
File est_CCI_expr_expr.tsv will save results of CCI through ligand and recepter expression.
}
\description{
Estimate three patterns for cell-cell interactions
}
