% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{ExprDistanceTable}
\alias{ExprDistanceTable}
\title{Save cell expression - distance patterns for simulation}
\usage{
ExprDistanceTable(
  gobject,
  in_hvg = F,
  output_file = file.path(getwd(), "est_CCI_dist_expr.csv"),
  region_specific = F,
  abs_log2fc_ICG = 0.25,
  p_adj = 0.05,
  spatial_network_name = "distance_based_network",
  seed = NULL
)
}
\arguments{
\item{gobject}{Giotto object, output of preprocessGiotto.}

\item{in_hvg}{Perform the analysis within the highly variable genes.}

\item{output_file}{Provide a path and file name to save the CCI results for use in sCCIgen.}

\item{region_specific}{Perform the analysis separately for each region.}

\item{abs_log2fc_ICG}{Effect size threshold for saving.}

\item{p_adj}{Adjusted p-value threshold (Default = 0.05) for saving.}

\item{spatial_network_name}{Networks to choose. Possible values include
"Delaunay_network", "distance_based_network", "knn_network".}

\item{seed}{Seed.}
}
\description{
Save cell expression - distance patterns for simulation
}
