% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{ExprExprTable}
\alias{ExprExprTable}
\title{Estimate cell expression -expression patterns and save the table for simulation}
\usage{
ExprExprTable(
  gobject,
  database = c("mouse", "human", "external"),
  external_database_path = NULL,
  region_specific = F,
  p_adj = 0.05,
  abs_log2fc_LR = 0.25,
  save_folder = getwd()
)
}
\arguments{
\item{expr_data}{Input expression data.}
}
\description{
Estimate cell expression -expression patterns and save the table for simulation
}
