% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{ExprExprTable_1region}
\alias{ExprExprTable_1region}
\title{Estimate  expression-expression patterns and save the table for simulation}
\usage{
ExprExprTable_1region(
  gobject,
  r,
  cell_meta,
  select_ligands,
  select_receptors,
  p_adj = 0.05,
  abs_log2fc_LR = 0.25
)
}
\arguments{
\item{p_adj}{Adjusted p-value threshold (Default = 0.05).}

\item{expr_data}{Input expression data.}

\item{spatial_data}{Input spatial data.}

\item{save_folder}{Provide a path to a folder that saves the CCI results for use in sCCIgen.}

\item{abs_enrichm}{Effect size threshold.}
}
\description{
Estimate  expression-expression patterns and save the table for simulation
}
