% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{generate_marginal_params}
\alias{generate_marginal_params}
\title{Generate model parameters from data.}
\usage{
generate_marginal_params(expr, feature, region = NULL, ncores = 1)
}
\arguments{
\item{expr}{Gene expression level (count).}

\item{feature}{Cell features (e.g. cell type, spatial coordinates, regions) of reference data.}

\item{ncores}{No of cores}

\item{region_specific_model}{Whether estimation model differ in different regions.}
}
\value{
Provide model parameters including marginal distributions and copula
(if not NULL) for all cell types in all regions.
}
\description{
This function generates model parameters from input data for all cell types in regions.
}
