% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{MergeRegion}
\alias{MergeRegion}
\title{Merge spatial and expression data from multiple regions}
\usage{
MergeRegion(points.list, expr.list)
}
\arguments{
\item{points.list}{points.list is a list of points from multiple regions}

\item{expr.list}{expr.list is a list of expressions from multiple regions}
}
\value{
\item{meta:}{Output the meta data of all cells in all regions}
\item{count:}{Output the count data of all cells in all regions}
}
\description{
Merge spatial and expression data from multiple regions
}
