% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCell.R
\name{multicell}
\alias{multicell}
\title{Assigns cells into spatial spots for multi-cell resolution data}
\usage{
multicell(expr, spatial, NoSpot = 500, cl = 1)
}
\arguments{
\item{expr}{expression profile of the cells}

\item{spatial}{spatial features like their spatial X and Y coordinates.}

\item{NoSpot}{Number of targeted spots.}

\item{cl}{No. of cores.}
}
\value{
\item{count:}{Expression profile of the spots.}
\item{spot_feature:}{Spot feature such as spot center coordinates and cell allocations
within the spots .}
}
\description{
This function assigns cells into spatial spots. Each spot may contain zero, one, or
multiple cells. Spots with zero cells won't be in the output.
}
