% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterDigest.R
\name{ParaCellsST}
\alias{ParaCellsST}
\title{Use parameters to simulate cell location based on modeling of existing SRT}
\usage{
ParaCellsST(para, spatial, seed_list)
}
\arguments{
\item{para}{Parameters loaded and cleaned from the parameter file using function
\code{ParaDigest}.}

\item{spatial}{Cell spatial data (cannot unpaired with expression data).}

\item{seed_list}{Seeds for all simulated data}
}
\value{
Simulated cell spatial map file.
}
\description{
Use parameters to simulate cell location. Here fit models on existing SRT data for simulation.
}
