% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprSimulator.R
\name{Pattern.Adj}
\alias{Pattern.Adj}
\title{Adjust the count data for all cells in all regions based on the
input spatial patterns}
\usage{
Pattern.Adj(
  sim.count,
  pattern.list = NULL,
  bond.extreme = T,
  keep.total.count = F,
  integer = T
)
}
\arguments{
\item{sim.count}{Spatial info for cell type 1 (e.g. neuron)}

\item{pattern.list}{A list of spatial patterns, which can be generated
from \code{Add.Spatial.Expr.Pattern}, \code{Add.Distance.Asso.Pattern}, or \code{Add.Expr.Asso.Pattern}.}

\item{bond.extreme}{Whether to bond extreme high values generated from
large effect sizes of spatial patterns (default = TRUE). If TRUE, no
gene can have more than the maximum of (1) 5 times the 97.5 percentile of the pattern adjusted
count and (2) the maximum of the original count.}

\item{keep.total.count}{If additional spatial patterns are added, whether
to rescale expression levels of all genes to keep the sequencing depth
(default = FALSE).}

\item{integer}{Whether to keep counts as integer (default=TRUE).}
}
\value{
Updated simulated gene expression counts by taking into consideration
of spatial patterns.
}
\description{
Adjust the count data for all regions based on the input spatial patterns
}
