% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{preprocessGiotto}
\alias{preprocessGiotto}
\title{Process SRT data using Giotto for CCI analysis.}
\usage{
preprocessGiotto(expr_data, loc, moreMeta)
}
\description{
This function takes input data and goes through data processing procedures for CCI analysis.
}
