% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactive_sCCIgen.R
\name{run_interactive_sCCIgen}
\alias{run_interactive_sCCIgen}
\title{Run interactive sCCIgen}
\usage{
run_interactive_sCCIgen()
}
\value{
A simulated dataset. Either individual csv files with counts
and cell features, or a Giotto object.
}
\description{
Run interactive sCCIgen
}
