% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{sCCIgen_to_Giotto}
\alias{sCCIgen_to_Giotto}
\title{Convert sCCIgen output to a Giotto object}
\usage{
sCCIgen_to_Giotto(parameter_file = "config.yml")
}
\arguments{
\item{parameter_file}{name for the config file. Default = config.yml}
}
\value{
a Giotto object
}
\description{
Convert sCCIgen output to a Giotto object
}
\examples{
giotto_object <- sCCIgen_to_Giotto()
}
