% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{sCCIgen_to_Seurat}
\alias{sCCIgen_to_Seurat}
\title{Convert sCCIgen output to a Seurat object}
\usage{
sCCIgen_to_Seurat(parameter_file = "config.yml")
}
\arguments{
\item{parameter_file}{name for the config file. Default = config.yml}
}
\value{
a Seurat object
}
\description{
Convert sCCIgen output to a Seurat object
}
\examples{
seurat_object <- sCCIgen_to_Seurat()
}
