% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{sCCIgen_to_SpatialExperiment}
\alias{sCCIgen_to_SpatialExperiment}
\title{Convert sCCIgen output to a SpatialExperiment object}
\usage{
sCCIgen_to_SpatialExperiment(parameter_file = "config.yml")
}
\arguments{
\item{parameter_file}{name for the config file. Default = config.yml}
}
\value{
a SpatialExperiment object
}
\description{
Convert sCCIgen output to a SpatialExperiment object
}
\examples{
spe_object <- sCCIgen_to_SpatialExperiment()
}
