% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{SpatialTable}
\alias{SpatialTable}
\title{SpatialTable}
\usage{
SpatialTable(
  gobject,
  top_num = 2,
  fdr_cut = 0.05,
  output_file = file.path(getwd(), "est_region_specific_genes.csv")
)
}
\arguments{
\item{gobject}{Giotto object.}

\item{top_num}{Keep top K number of genes for each cell type}

\item{fdr_cut}{Keep genes whose FDR is less than this cutoff.}

\item{output_file}{output_file}
}
\description{
SpatialTable
}
