% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointSimulator_STData.R
\name{cell.loc.model.fc}
\alias{cell.loc.model.fc}
\title{Generate cell location data by modeling the spatial information of existing data.}
\usage{
cell.loc.model.fc(
  n,
  PointLoc,
  PointAnno,
  PointRegion,
  window_method,
  seed = NULL,
  cell.inh.attr.input1 = NULL,
  same.dis.cutoff = 0.01,
  even.distribution.coef = 0
)
}
\arguments{
\item{n}{No. of cells}

\item{PointLoc}{The location of input cells on x, y axis.}

\item{PointAnno}{The cell type annotation of input cells.}

\item{PointRegion}{The spatial regions of input cells.}

\item{window_method}{Method for estimating window of cells. Choices include "network", "convex", "convex2",
"convex3", convex5", "rectangle". Default is "network", which is most accurate.}

\item{seed}{Random seed.}

\item{same.dis.cutoff}{Cutoff value for regarding two cells are in the same location (default = 0).}

\item{even.distribution.coef}{A parameter to avoid cells gathering from each other.}
}
\value{
Simulated cell spatial map file.
}
\description{
Poisson random process model separately for each cell type is the foundation. The impact of CCIs and other spatial patterns update the
initial cell allocation.
}
