% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{cellProximityTable}
\alias{cellProximityTable}
\title{Estimate cell attraction and inhibition patterns and save the table for simulation}
\usage{
cellProximityTable(
  gobject,
  output_file = file.path(getwd(), "est_CCI_dist_dist.csv"),
  abs_enrichm = 0.3,
  p_adj = 0.05,
  spatial_network_name = "Delaunay_network",
  save.unfiltered = T,
  seed = NULL
)
}
\arguments{
\item{output_file}{Provide a path and file name to save the CCI results for use in sCCIgen.}

\item{abs_enrichm}{Effect size threshold.}

\item{p_adj}{Adjusted p-value threshold (Default = 0.05).
File est_CCI_dist_dist.tsv will save results of CCI through cell location attraction and inhibition.
File est_CCI_expr_dist.tsv will save results of CCI through expression via cell location impact.
File est_CCI_expr_expr.tsv will save results of CCI through ligand and recepter expression.}

\item{expr_data}{Input expression data.}

\item{spatial_data}{Input spatial data.}
}
\description{
Estimate cell attraction and inhibition patterns and save the table for simulation
}
