% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{ExprDistanceTable}
\alias{ExprDistanceTable}
\title{Estimate cell expression - distance patterns and save the table for simulation}
\usage{
ExprDistanceTable(
  gobject,
  in_hvg = F,
  save_folder = getwd(),
  region_specific = F,
  abs_log2fc_ICG = 0.25,
  p_adj = 0.05
)
}
\arguments{
\item{expr_data}{Input expression data.}
}
\description{
Estimate cell expression - distance patterns and save the table for simulation
}
