% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{ExprExprTable}
\alias{ExprExprTable}
\title{Save cell expression - cell expression association patterns for simulation}
\usage{
ExprExprTable(
  gobject,
  database = c("mouse", "human", "external"),
  external_database_path = NULL,
  region_specific = F,
  spatial_network_name,
  p_adj = 0.05,
  abs_log2fc_LR = 0.25,
  direction = c("both", "positive", "negative"),
  output_file = file.path(getwd(), "est_CCI_expr_expr.csv"),
  seed = NULL
)
}
\arguments{
\item{gobject}{Giotto object, output of preprocessGiotto.}

\item{database}{Specify databases for pairs of genes for consideration. Possible values include
"mouse", "human", and "external". The "mouse" and "human" are ligand-receptor pairs downloaded from CellTalkDB.
If "external" is used, users also need to provide value for \code{external_database_path}.}

\item{external_database_path}{A path and file name to external gene-gene pair database.}

\item{region_specific}{Perform the analysis separately for each region.}

\item{spatial_network_name}{Networks to choose. Possible values include
"Delaunay_network", "distance_based_network", "knn_network".}

\item{p_adj}{Adjusted p-value threshold (Default = 0.05) for saving.}

\item{abs_log2fc_LR}{Effect size threshold for saving.}

\item{direction}{Keep positive, negative, or both associations.}

\item{output_file}{Provide a path and file name to save the CCI results for use in sCCIgen.}

\item{seed}{Seed.}
}
\description{
Save cell expression - cell expression association patterns for simulation
}
