% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{ExprExprTable}
\alias{ExprExprTable}
\title{Estimate cell expression -expression patterns and save the table for simulation}
\usage{
ExprExprTable(
  gobject,
  database = c("mouse", "human", "external"),
  external_database_path = NULL,
  region_specific = F,
  spatial_network_name,
  p_adj = 0.05,
  abs_log2fc_LR = 0.25,
  output_file = file.path(getwd(), "est_CCI_expr_expr.csv"),
  seed = NULL
)
}
\arguments{
\item{expr_data}{Input expression data.}
}
\description{
Estimate cell expression -expression patterns and save the table for simulation
}
