% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCCI.R
\name{preprocessGiotto}
\alias{preprocessGiotto}
\title{Process SRT data using Giotto for CCI analysis.}
\usage{
preprocessGiotto(
  expr_data,
  spatial_data,
  run_hvg = T,
  run_kNN_network = F,
  run_Delaunay_network = F,
  run_Dist_network = F,
  dis.cut = NULL
)
}
\description{
This function takes input data and goes through data processing procedures for CCI analysis.
}
\details{
#' @import Giotto
}
