% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactive_sCCIgen.R
\name{run_interactive_sCCIgen}
\alias{run_interactive_sCCIgen}
\title{Run interactive sCCIgen}
\usage{
run_interactive_sCCIgen()
}
\value{
Either a .yml parameter file or a simulated dataset. If running the
simulation, individual .csv files with counts and cell features are locally
saved. Optionally, a Giotto, Seurat, or SpatialExperiment object is created
and saved locally as an .RDS file.
}
\description{
Run interactive sCCIgen
}
