% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_interactive_STsimulator.R
\name{run_interactive_sCCIgen}
\alias{run_interactive_sCCIgen}
\title{Run interactive STsimulator}
\usage{
run_interactive_sCCIgen()
}
\value{
A simulated dataset. Either as individual csv files with counts, cell features, or a Giotto object.
}
\description{
Run interactive STsimulator
}
