% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{sCCIgen_to_SpatialExperiment}
\alias{sCCIgen_to_SpatialExperiment}
\title{Convert sCCIgen output to a SpatialExperiment object}
\usage{
sCCIgen_to_SpatialExperiment(counts_file, metadata_file)
}
\arguments{
\item{counts_file}{path to .tsv file with gene x cell counts matrix.}

\item{metadata_file}{path to .tsv file with metadata. The file should
contain the following columns: Cell, annotation, x.loc, y.loc, and region.}
}
\value{
a SpatialExperiment object
}
\description{
Convert sCCIgen output to a SpatialExperiment object
}
\examples{
spe_object <- sCCIgen_to_SpatialExperiment()
}
